+++ src/i_input.cpp
@@ -422,7 +422,7 @@ static int AccelerateMouse(int val)
 
     if (val > mouse_threshold)
     {
-        return static_cast<int>((val - mouse_threshold) * mouse_acceleration + mouse_threshold);
+        return static_cast<int>((static_cast<float>(val - mouse_threshold)) * mouse_acceleration + static_cast<float>(mouse_threshold));
     }
     else
     {
@@ -438,7 +438,7 @@ static int AccelerateMouseY(int val)
 
     if (val > mouse_threshold_y)
     {
-        return static_cast<int>((val - mouse_threshold_y) * mouse_acceleration_y + mouse_threshold_y);
+        return static_cast<int>((static_cast<float>(val - mouse_threshold_y)) * mouse_acceleration_y + static_cast<float>(mouse_threshold_y));
     }
     else
     {
