+++ src/net_server.cpp
@@ -1172,7 +1172,7 @@ static void NET_SV_ParseGameData(net_packet_t *packet, net_client_t *client)
             return;
         }
 
-        index = seq + i - recvwindow_start;
+        index = static_cast<int>(seq + i - recvwindow_start);
 
         if (index < 0 || index >= BACKUPTICS)
         {
